+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Adds `MapX.NewEntityFn`, `MapX.AddFn`, `Map.NewEntityFn` and `Map.AddFn` (#145)
 - Adds `ExchangeX.AddFn` and `ExchangeX.ExchangeFn` (#145)
 - Adds `World.Stats` for extracting world statistics (#147)
+- Adds `Unsafe.IDs` to get all component IDs of an entity (#149)
 
 ### Documentation